// src/interfaces/currency.interface.ts
export interface Products {
  products: Product[];
}

interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}

export interface DiscountedProducts {
  products: DiscountedProduct[];
}

interface DiscountedProduct {
  title: string;
  price: number;
}